requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
bleach==6.1.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed_rtf==1.0.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
DataProperty==1.0.1
datasets==3.0.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
easygui==0.98.3
ebcdic==1.1.1
einops==0.8.0
extract-msg==0.49.0
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
flask-cloudflared==0.0.14
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.1.9
llvmlite==0.42.0
lm-eval==0.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mbstrdecoder==1.1.3
mdurl==0.1.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.9.2
networkx==3.3
nltk==3.9.1
numba==0.59.1
numexpr==2.10.1
numpy==1.26.4
olefile==0.47
oletools==0.60.2
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
optimum==1.17.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pathvalidate==3.2.1
pcodedmp==1.2.6
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
peft==0.12.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
protobuf==5.28.2
psutil==6.0.0
pyarrow==17.0.0
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
pypdfium2==4.30.0
pyreadline3==3.5.4
pytablewriter==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
pywin32==306
PyYAML==6.0.2
red-black-tree-mod==1.20
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
rich==13.9.2
rouge_score==0.1.2
rpds-py==0.20.0
RTFDE==0.1.2
ruff==0.6.9
sacrebleu==1.5.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.16.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
sqlitedict==2.1.0
sse-starlette==1.6.5
starlette==0.38.6
sympy==1.13.3
tabledata==1.3.3
tabula==1.0.5
tcolorpy==0.1.6
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
tomlkit==0.12.0
torch==2.4.1
tqdm==4.66.5
tqdm-multiprocess==0.0.11
transformers==4.45.2
typepy==1.3.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.1
wandb==0.18.3
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.44.0
win_unicode_console==0.5
xxhash==3.5.0
yarl==1.13.1
zstandard==0.23.0
